+++ database/user/mysql_user.sql
@@ -1,0 +1,3 @@
+CREATE USER 'quicksight_readonly'@'%' IDENTIFIED BY 'my_pass';
+GRANT SELECT ON backend2024.* TO 'quicksight_readonly'@'%';
+FLUSH PRIVILEGES;