+++ sql/_test.sql
@@ -1,1 +1,0 @@
-SELECT * FROM live.strategy_instances LIMIT 1