+++ sql/data/clean_exposure.sql
@@ -1,0 +1,1 @@
+DELETE FROM quicksight._exposure WHERE trading_minute<CURRENT_DATE - Interval '8 day'