+++ sql/data/clean_tradingdata.sql
@@ -1,0 +1,1 @@
+DELETE FROM quicksight._tradingdata WHERE trading_minute<CURRENT_DATE - Interval '8 day'