+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Unreleased]
 ### Added
+- Added intraday views and tables
+  - `quicksight.intraday`
+  - `quicksight.intraday_by_exchange`
+  - `quicksight.intraday_by_account`
+  - `quicksight.intraday_top_instruments`
+- Added deploy script to deploy to monkey-bi server
+  - `composer deploy`
+  - `composer dev`
 - Added user scripts to `database/user`
   - scripts
     - `create_users.sql`
@@ -10,6 +18,7 @@
 - Added `database/permission.sql` that defines all permissions
 - Added generic SQL-Script-Runner
 ### Changed
+- Need `sswrockfunds/core 0.15.0` as min dependency
 ### Fixed
 ### Removed
 ### -